requirements.txt
langchain == 0.0.154
PyPDF -- 3.0.1
python-dotenv==1.0.0
stramlit == 1.18.1
faiss-cpu==1.7.4
streamlit-extras